(PC-7631) Make Home the initial screen in order to avoid a blank screen on back navigation
